Auth para rutas

+++ routes/api.php
@@ -28,53 +28,52 @@ Route::middleware('auth:sanctum')->get('/user', function (Request $request) {
 Route::post('/login', [SuperAdminAuthController::class, 'login']);
 Route::post('/logout', [SuperAdminAuthController::class, 'logout'])->middleware('auth:superadmin');
 
-Route::get('/superAdmins', [SuperAdminController::class, 'index']);
-Route::post('/superAdmins', [SuperAdminController::class, 'store']);
-Route::get('/superAdmins/{superAdmin}', [SuperAdminController::class, 'show']);
-Route::put('/superAdmins/{superAdmin}', [SuperAdminController::class, 'update']);
-Route::delete('/superAdmins/{superAdmin}', [SuperAdminController::class, 'destroy']);
+Route::get('/superAdmins', [SuperAdminController::class, 'index'])->middleware('auth');
+Route::post('/superAdmins', [SuperAdminController::class, 'store'])->middleware('auth');
+Route::get('/superAdmins/{superAdmin}', [SuperAdminController::class, 'show'])->middleware('auth');
+Route::put('/superAdmins/{superAdmin}', [SuperAdminController::class, 'update'])->middleware('auth');
+Route::delete('/superAdmins/{superAdmin}', [SuperAdminController::class, 'destroy'])->middleware('auth');
 
-Route::get('/promotores', [PromotorController::class, 'index']);
-Route::post('/promotores', [PromotorController::class, 'store']);
-Route::get('/promotores/{promotor}', [PromotorController::class, 'show']);
-Route::put('/promotores/{promotor}', [PromotorController::class, 'update']);
-Route::delete('/promotores/{promotor}', [PromotorController::class, 'destroy']);
-Route::get('/promotor/{promotorId}/promoteds', [PromotorController::class, 'showPromoteds']);
-Route::get('/promotor/{promotorId}/promoteds/count', [PromotorController::class, 'showPromotedsCount']);
-
-
-Route::get('/promoted', [PromotedController::class, 'index']);
-Route::post('/promoted', [PromotedController::class, 'store']);
-Route::get('/promoted/{id}', [PromotedController::class, 'show']);
-Route::put('/promoted/{id}', [PromotedController::class, 'update']);
-Route::delete('/promoted/{id}', [PromotedController::class, 'destroy']);
+Route::get('/promotores', [PromotorController::class, 'index'])->middleware('auth');
+Route::post('/promotores', [PromotorController::class, 'store'])->middleware('auth');
+Route::get('/promotores/{promotor}', [PromotorController::class, 'show'])->middleware('auth');
+Route::put('/promotores/{promotor}', [PromotorController::class, 'update'])->middleware('auth');
+Route::delete('/promotores/{promotor}', [PromotorController::class, 'destroy'])->middleware('auth');
+Route::get('/promotor/{promotorId}/promoteds', [PromotorController::class, 'showPromoteds'])->middleware('auth');
+Route::get('/promotor/{promotorId}/promoteds/count', [PromotorController::class, 'showPromotedsCount'])->middleware('auth');
 
 
-
-
-Route::post('/upload-excel/{promotorId}', [PromotedController::class, 'uploadExcel']);
-Route::get('/export-excel', [PromotedController::class, 'export']);
-
-Route::get('/municipals', [MunicipalController::class, 'index']);
-Route::get('/municipals/{municipalId}/districts/{districtId}/promoveds/count', [MunicipalController::class, 'countPromovedsInDistrict']);
-Route::get('/municipals/districts/promoveds/count', [MunicipalController::class, 'countPromovedsInAll']);
-Route::get('/municipals/sections/promoveds/count', [MunicipalController::class, 'countPromovedsInSections']);
-Route::get('/municipals/sections/promoveds/count-by-date', [MunicipalController::class, 'countPromovedsInSectionsByDate']);
-Route::get('/municipals/districts/promoveds/count-by-date', [MunicipalController::class, 'countPromovedsInDistrictsByDate']);
-Route::get('/municipals/sections-with-promoved-count', [MunicipalController::class, 'sectionsWithPromovedCount']);
-
-
-Route::get('/districts/{id}', [DistrictController::class, 'show']);
-
-
-Route::get('/problems', [ProblemController::class, 'index']); // Listar todos los problemas
-Route::post('/problems', [ProblemController::class, 'store']); // Crear un nuevo problema
-Route::get('/problems/{problem}', [ProblemController::class, 'show']); // Mostrar un problema específico
-Route::put('/problems/{problem}', [ProblemController::class, 'update']); // Actualizar un problema
-Route::delete('/problems/{problem}', [ProblemController::class, 'destroy']); // Eliminar un problema
+Route::get('/promoted', [PromotedController::class, 'index'])->middleware('auth');
+Route::post('/promoted', [PromotedController::class, 'store'])->middleware('auth');
+Route::get('/promoted/{id}', [PromotedController::class, 'show'])->middleware('auth');
+Route::put('/promoted/{id}', [PromotedController::class, 'update'])->middleware('auth');
+Route::delete('/promoted/{id}', [PromotedController::class, 'destroy'])->middleware('auth');
 
 
 
-Route::get('/catalog/municipal', [CatalogController::class, 'MunicipalSelect']);
-Route::get('/municipals/{municipalId}/districts', [CatalogController::class, 'getDistrictsByMunicipal']);
-Route::get('/districts/{districtId}/sections', [CatalogController::class, 'getSectionsByDistrict']);
+
+Route::post('/upload-excel/{promotorId}', [PromotedController::class, 'uploadExcel'])->middleware('auth');
+Route::get('/export-excel', [PromotedController::class, 'export'])->middleware('auth');
+
+Route::get('/municipals', [MunicipalController::class, 'index'])->middleware('auth');
+Route::get('/municipals/{municipalId}/districts/{districtId}/promoveds/count', [MunicipalController::class, 'countPromovedsInDistrict'])->middleware('auth');
+Route::get('/municipals/districts/promoveds/count', [MunicipalController::class, 'countPromovedsInAll'])->middleware('auth');
+Route::get('/municipals/sections/promoveds/count', [MunicipalController::class, 'countPromovedsInSections'])->middleware('auth');
+Route::get('/municipals/sections/promoveds/count-by-date', [MunicipalController::class, 'countPromovedsInSectionsByDate'])->middleware('auth');
+Route::get('/municipals/districts/promoveds/count-by-date', [MunicipalController::class, 'countPromovedsInDistrictsByDate'])->middleware('auth');
+Route::get('/municipals/sections-with-promoved-count', [MunicipalController::class, 'sectionsWithPromovedCount'])->middleware('auth');
+
+
+Route::get('/districts/{id}', [DistrictController::class, 'show'])->middleware('auth');
+
+Route::get('/problems', [ProblemController::class, 'index'])->middleware('auth'); // Listar todos los problemas
+Route::post('/problems', [ProblemController::class, 'store'])->middleware('auth'); // Crear un nuevo problema
+Route::get('/problems/{problem}', [ProblemController::class, 'show'])->middleware('auth'); // Mostrar un problema específico
+Route::put('/problems/{problem}', [ProblemController::class, 'update'])->middleware('auth'); // Actualizar un problema
+Route::delete('/problems/{problem}', [ProblemController::class, 'destroy'])->middleware('auth'); // Eliminar un problema
+
+
+
+Route::get('/catalog/municipal', [CatalogController::class, 'MunicipalSelect'])->middleware('auth');
+Route::get('/municipals/{municipalId}/districts', [CatalogController::class, 'getDistrictsByMunicipal'])->middleware('auth');
+Route::get('/districts/{districtId}/sections', [CatalogController::class, 'getSectionsByDistrict'])->middleware('auth');
